tested task added treadpool (stolen on net)

+++ core/src/general/System.hpp
@@ -6,7 +6,7 @@ using namespace std::chrono;
 class System {
 
 public:
-	inline static long currentTimeMillis() {
+	inline static long long currentTimeMillis() {
 		return duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count();
 	}
 };
